MAKE WEBCIT GREAT AGAIN!

+++ webcit-ng/admin_functions.c
@@ -1,7 +1,7 @@
 /*
  * Admin functions
  *
- * Copyright (c) 1996-2017 by the citadel.org team
+ * Copyright (c) 1996-2018 by the citadel.org team
  *
  * This program is open source software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3.
@@ -32,7 +32,7 @@ void try_login(struct http_transaction *h, struct ctdlsession *c)
 	snprintf(buf, sizeof buf, "%s:%s", username, password);
 	CtdlEncodeBase64(auth, buf, strlen(buf), 0);
 
-	syslog(LOG_DEBUG, "try_login(username='%s',password=(%d bytes))", username, strlen(password));
+	syslog(LOG_DEBUG, "try_login(username='%s',password=(%d bytes))", username, (int)strlen(password));
 
 	ctdl_printf(c, "LOUT");							// log out, in case we were logged in
 	ctdl_readline(c, buf, sizeof(buf));					// ignore the result
