+++ webcit-ng/messages.c
@@ -1,7 +1,7 @@
 /*
  * Message base functions
  *
- * Copyright (c) 1996-2016 by the citadel.org team
+ * Copyright (c) 1996-2018 by the citadel.org team
  *
  * This program is open source software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3.
@@ -269,7 +269,7 @@ void download_mime_component(struct http_transaction *h, struct ctdlsession *c, 
 	do {
 		thisblock = read(c->sock, &h->response_body[bytes], (h->response_body_length - bytes));
 		bytes += thisblock;
-		syslog(LOG_DEBUG, "Bytes read: %d of %d", bytes, h->response_body_length);
+		syslog(LOG_DEBUG, "Bytes read: %d of %d", (int)bytes, (int)h->response_body_length);
 	} while ((bytes < h->response_body_length) && (thisblock >= 0));
 	h->response_body[h->response_body_length] = 0;				// null terminate it just for good measure
 	syslog(LOG_DEBUG, "content type: %s", content_type);
